Fix pending tx functionality + expose the correct interface
