+++ ethereumj-core/src/main/java/org/ethereum/manager/WorldManager.java
@@ -71,7 +71,6 @@ public class WorldManager {
     @Autowired
     private AdminInfo adminInfo;
 
-    private final Set<Transaction> pendingTransactions = Collections.synchronizedSet(new HashSet<Transaction>());
 
     @Autowired
     private EthereumListener listener;
@@ -99,22 +98,6 @@ public class WorldManager {
     public void stopPeerDiscovery() {
         if (peerDiscovery.isStarted())
             peerDiscovery.stop();
-    }
-
-    public void addPendingTransactions(Set<Transaction> transactions) {
-        logger.info("Pending transaction list added: size: [{}]", transactions.size());
-
-        if (listener != null)
-            listener.onPendingTransactionsReceived(transactions);
-        pendingTransactions.addAll(transactions);
-    }
-
-    public void clearPendingTransactions(List<Transaction> receivedTransactions) {
-
-        for (Transaction tx : receivedTransactions) {
-            logger.info("Clear transaction, hash: [{}]", Hex.toHexString(tx.getHash()));
-            pendingTransactions.remove(tx);
-        }
     }
 
     public ChannelManager getChannelManager() {
@@ -153,14 +136,11 @@ public class WorldManager {
         return activePeer;
     }
 
-    public Set<Transaction> getPendingTransactions() {
-        return pendingTransactions;
-    }
 
     public boolean isBlockchainLoading() {
         return blockchain.getQueue().size() > 2;
     }
-    
+
     public void loadBlockchain() {
 
         Block bestBlock = blockStore.getBestBlock();
@@ -176,6 +156,7 @@ public class WorldManager {
             blockchain.setBestBlock(Genesis.getInstance());
             blockchain.setTotalDifficulty(BigInteger.ZERO);
 
+            listener.onBlock(Genesis.getInstance());
             repository.dumpState(Genesis.getInstance(), 0, 0, null);
 
             logger.info("Genesis block loaded");
